Backend/requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langgraph
langchain-google-genai
langchain-groq
python-dotenv
pydantic
langchain-groq
